autoping clean (#2289)

+++ src/main/java/ti4/map/GameProperties.java
@@ -1,14 +1,14 @@
 package ti4.map;
 
-import lombok.Getter;
-import lombok.Setter;
-import ti4.helpers.Constants;
-
 import java.sql.Timestamp;
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+
+import lombok.Getter;
+import lombok.Setter;
+import ti4.helpers.Constants;
 
 @Getter
 @Setter
@@ -89,7 +89,7 @@ public class GameProperties {
     private String hexBorderStyle = "off";
     private String textSize = "medium";
     private String outputVerbosity = Constants.VERBOSITY_VERBOSE;
-    private long autoPingSpacer;
+    private int autoPingSpacer;
     private List<String> tags = new ArrayList<>();
 
     // Game modes / homebrew flags
